[PATCH] possible fix for bridge-hosts util

+++ utils/bridge-hosts.sh
@@ -106,38 +106,43 @@ BRIDGE_VERBOSE_FLAG=${BRIDGE_VERBOSE_FLAG_INIT:-}
 if [ "${DEBUG:-0}" -eq 1 ]; then BRIDGE_VERBOSE_FLAG="-v"; fi;
 
 # TODO: verify ANYVM_BRIDGE_HOSTS_FILE is a file that exists
+if [ -f "${ANYVM_BRIDGE_HOSTS_FILE:-}" ]; then
 
-debug_sub_log "Preparing script to bridge hosts on Guest VM" ;
-BRIDGE_HOSTS_SCRIPT_PATH="$BRIDGE_DATA_DIR/bridge_hosts_$$.sh"
-cp ${BRIDGE_VERBOSE_FLAG:-} -f "${ANYVM_BRIDGE_HOSTS_FILE}" "$BRIDGE_HOSTS_SCRIPT_PATH"
-debug_sub_log "=> Staged" & debug_sub_log "..=> Setting Permissions on staged script" ;
-chmod ${BRIDGE_VERBOSE_FLAG:-} +x "$BRIDGE_HOSTS_SCRIPT_PATH"
+  debug_sub_log "Preparing script to bridge hosts on Guest VM" ;
+  BRIDGE_HOSTS_SCRIPT_PATH="$BRIDGE_DATA_DIR/bridge_hosts_$$.sh"
+  cp ${BRIDGE_VERBOSE_FLAG:-} -f "${ANYVM_BRIDGE_HOSTS_FILE}" "$BRIDGE_HOSTS_SCRIPT_PATH"
+  debug_sub_log "=> Staged" & debug_sub_log "..=> Setting Permissions on staged script" ;
+  chmod ${BRIDGE_VERBOSE_FLAG:-} +x "$BRIDGE_HOSTS_SCRIPT_PATH"
 
-debug_sub_log "Ready to transfer \"${BRIDGE_HOSTS_SCRIPT_PATH}\" to Guest VM" ;
+  debug_sub_log "Ready to transfer \"${BRIDGE_HOSTS_SCRIPT_PATH}\" to Guest VM" ;
 
-debug_sub_log "=> Using runner /etc/hosts data to bridge hosts on Guest VM" ;
-BRIDGE_HOSTS_DATA_PATH="$BRIDGE_DATA_DIR/hosts_$$.data"
+  debug_sub_log "=> Using runner /etc/hosts data to bridge hosts on Guest VM" ;
+  BRIDGE_HOSTS_DATA_PATH="$BRIDGE_DATA_DIR/hosts_$$.data"
 
-# copy host file to guest /tmp/hosts.from_host
-if [ -f /etc/hosts ]; then
-  cat <"/etc/hosts" >> "$BRIDGE_HOSTS_SCRIPT_PATH" ; # 'copy' but not permissions (by reading)
-  debug_sub_log "..=> Data Staged"
-  debug_sub_log "=> Ready to also transfer \"${BRIDGE_HOSTS_DATA_PATH}\" to Guest VM" ;
-  debug_sub_log "....=> Waiting for transfer" ;
-  scp $SSH_EPHEMERAL_OPTS -P ${BRIDGE_VM_PORT:-22} "${BRIDGE_HOSTS_DATA_PATH}" root@"$BRIDGE_VM":/tmp/hosts.from_host || printf '::warning:: %s\n' "failed to scp bridge-hosts data"
-  scp $SSH_EPHEMERAL_OPTS -P $BRIDGE_VM_PORT "$BRIDGE_HOSTS_SCRIPT_PATH" root@"$BRIDGE_VM":/tmp/bridge-hosts.sh || printf '::warning:: %s\n' "failed to scp bridge-hosts script"
-  debug_sub_log "..=> Transferred" & debug_sub_log "..=> Waiting for bridging" &
-  # remote merge script: run on guest (idempotent-ish)
-  ssh $SSH_EPHEMERAL_OPTS -p ${BRIDGE_VM_PORT:-22} root@"$BRIDGE_VM" "sh /tmp/bridge-hosts.sh" || printf '::error:: %s\n' "warning: bridge-hosts execution failed"
-  rm -f "$BRIDGE_HOSTS_DATA_PATH" 2>/dev/null || true ;
-  debug_sub_log "=> Bridged"
-else
-  printf '::warning:: %s\n' "/etc/hosts not found locally; nothing to merge." >&2
-  debug_sub_log "Nothing transferred"
-fi
+  # copy host file to guest /tmp/hosts.from_host
+  if [ -f /etc/hosts ]; then
+    cat <"/etc/hosts" >> "$BRIDGE_HOSTS_DATA_PATH" ; # 'copy' but not permissions (by reading)
+    debug_sub_log "..=> Data Staged"
+    debug_sub_log "=> Ready to also transfer \"${BRIDGE_HOSTS_DATA_PATH}\" to Guest VM" ;
+    debug_sub_log "....=> Waiting for transfer" ;
+    scp $SSH_EPHEMERAL_OPTS -P ${BRIDGE_VM_PORT:-22} "${BRIDGE_HOSTS_DATA_PATH}" root@"$BRIDGE_VM":/tmp/hosts.from_host || printf '::warning:: %s\n' "failed to scp bridge-hosts data"
+    scp $SSH_EPHEMERAL_OPTS -P $BRIDGE_VM_PORT "$BRIDGE_HOSTS_SCRIPT_PATH" root@"$BRIDGE_VM":/tmp/bridge-hosts.sh || printf '::warning:: %s\n' "failed to scp bridge-hosts script"
+    debug_sub_log "..=> Transferred" & debug_sub_log "..=> Waiting for bridging" &
+    # remote merge script: run on guest (idempotent-ish)
+    ssh $SSH_EPHEMERAL_OPTS -p ${BRIDGE_VM_PORT:-22} root@"$BRIDGE_VM" "sh /tmp/bridge-hosts.sh" || printf '::error:: %s\n' "warning: bridge-hosts execution failed"
+    rm -f "$BRIDGE_HOSTS_DATA_PATH" 2>/dev/null || true ;
+    debug_sub_log "=> Bridged"
+  else
+    printf '::warning:: %s\n' "/etc/hosts not found locally; nothing to merge." >&2
+    debug_sub_log "Nothing transferred"
+  fi
 
-# best effort cleanup
+  # best effort cleanup
 rm ${BRIDGE_VERBOSE_FLAG:-} -f "$BRIDGE_HOSTS_SCRIPT_PATH}" 2>/dev/null || true ; # un-stage as needed (but never error)
+else
+  printf '::error:: %s\n' "Error missing expected resource at '${ANYVM_BRIDGE_HOSTS_FILE:-}'" ;
+fi ;
+
 unset BRIDGE_VM
 unset BRIDGE_VM_PORT
 unset BRIDGE_DATA_DIR
